requirements.txt
torch
rembg
pillow
trimesh
matplotlib
pyrender
transformers
diffusers
